refactor(DebugConsole): Restructure code with mixins for better maintainability
