+++ app/components/DebugConsole/constants/logStyles.js
@@ -1,0 +1,12 @@
+//@license magnet:?xt=urn:btih:1f739d935676111cfff4b4693e3816e664797050&dn=gpl-3.0.txt GPL-v3
+
+import { Color } from "@nativescript/core/color";
+
+export const LOG_COLORS = {
+  HIGHLIGHT_BG: new Color("rgba(255, 255, 255, 0.15)"),
+  TRANSPARENT_BG: new Color("transparent")
+};
+
+export default {
+  LOG_COLORS
+};
